fixed save GAME COMPLETEDDD

+++ Assets/Scripts/Player.cs
@@ -22,7 +22,7 @@ public class Player : Damageable
     [SerializeField] private float CurrentExp;
     [SerializeField] private int LevelUpEXP;
     [SerializeField] public static float PlayerLevel;
-    public float PlayerLevelMainMenu;
+    public int PlayerLevelMainMenu;
     [SerializeField] public float PreviousLevel;
     [SerializeField] private float levelUpRequirementMultiplier = 1.1f;
 
@@ -113,7 +113,7 @@ public class Player : Damageable
     }
     public override void Death()
     {
-        PlayerLevelMainMenu = PlayerLevel;
+        PlayerLevelMainMenu = ((int)PlayerLevel);
         SoundFXManager.instance.PlaySoundFXClip(playerDeathClip, transform, 1f);
         GamesManager.Instance.SwitchState<GameOverState>();
     }
